# Optimize ingester Push() on blocks storage when the per-user or per-metric series limit is reached